Fixup imports, revert trash __init__
